requirements.txt
cbor2==5.4.6
click==8.1.3
evdev==1.6.1
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
luma.core==2.4.0
luma.led-matrix==1.7.0
MarkupSafe==2.1.2
Pillow==9.4.0
pkg-resources==0.0.0
pyftdi==0.54.0
pyserial==3.5
python-vlc==3.0.18121
pyusb==1.2.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
smbus2==0.4.2
spidev==3.6
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
